==> 03_render.Rmd <==
---
title: "Skills in the BC labour market"
author: "Ermias Afeworki, Richard Martin"
output:
  bookdown::word_document2:
    fig_caption: true
    fig_caption_location: bottom
    toc: false           # optional: add if you want a TOC
    number_sections: false # optional: number sections or not
    reference_docx: null  # optional: specify your Word template here
---

```{r, include=FALSE}
knitr::opts_chunk$set(
  echo = FALSE,
  warning = FALSE,
  message = FALSE,
  error = FALSE
)

if (file.info(here::here("out","richs_objects.rds"))$mtime < file.info("01_source_rich.R")$mtime) {
  stop("richs_objects.rds is stale — rerun 01_source_rich.R")
}
library(tidyverse)
ro <- readr::read_rds(here::here("out", "richs_objects.rds"))

all_occs_chart <- read_rds(here::here("out", "all_occs_chart.rds"))
plot_upper <- read_rds( here::here("out", "plot_upper.rds"))
plot_lower <- read_rds(here::here("out", "plot_lower.rds"))
plot_dist <- read_rds(here::here("out", "plot_dist.rds"))

library(readxl)
library(here)

# Read a specific worksheet
skill_clusters <- read_excel(here("out", "skill_clusters_data.xlsx"), sheet = "skill_clusters_tbl") |> 
  mutate(across(everything(), ~ ifelse(is.na(.), "", .))) # Replace NA with blank space


```

`r format(lubridate::now(), "%Y-%m-%d %H:%M:%S")#timestamp handy while developing`

In order for the BC economy to thrive it is essential that the skills of the workforce match the needs of employers. However, the labour market is constantly evolving both in terms of occupational composition as well as the skills required for each specific occupation. In this note we focus exclusively on composition changes, and make the assumption that occupational skill profiles are constant. This assumption is supported by data frequency: employment by occupation is available (and variable) monthly in the Labour Force Survey, whereas the O\*NET^[While O\*NET provides both importance and level ratings for skills and competencies, we focus on importance ratings in this report, as the two measures are highly correlated.]  skills data uses a rolling update schedule, completing a full refresh of all occupations every five years.

The goal is to identify the skills that are becoming more important in the labour market. We begin by quantifying the broad shifts in the labour market by looking at how the TEER composition has changed historically, and is forecast to change over the 10 years covered by the Labour Market Outlook. Once we have established the broad shifts in the labour market, we then compare the current skill profiles of High Opportunity Occupations to the rest of the labour market. Finally we look at how the distribution of skills is changing over time, both historically (2015-2024) and in the future (2025-2035).

## A first look: breakdown by TEER

We can get a preliminary idea of the changes in the labour market by looking at the breakdown of employment by TEER (Training, Education, Experience and Responsibilities) categories. The TEER system is a classification system used in Canada to categorize occupations based on the training, education, experience, and responsibilities required for each job. If the TEER composition of the labour market is changing over time it is a clear indication that the required skills are also changing.

```{r, fig1, fig.cap="Employment by TEER groupings", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
ro$emp_teer_plt 
```

## High opportunity skills

In recognition of the dynamic nature of the labour market, the Labour Market Outlook identifies certain occupations as being High Opportunity Occupations on the basis of seven economic indicators. These indicators can be grouped into four categories, which are based on current and future labour market conditions. Each occupation is given a score based on its ranking in each of the seven economic indicators, and scores are weighted based on their labour market impact. The occupations are then sorted by score, and the 125 occupations with the highest scores become High Opportunity Occupations.

In what follows we create two weighted average skill profiles, one for the High Opportunity Occupations and one for the rest of the labour market (Figure \@ref(fig:fig2)). It is evident that High Opportunity Occupations require higher skills, with a few exceptions e.g. repairing and equipment maintenance. Using these weighted average skill scores, we investigate the relative (Figure \@ref(fig:fig3)) and absolute (Figure \@ref(fig:fig4)) differences to identify high opportunity skills: the skills that have both relatively and absolutely higher scores in high opportunity occupations.  From these plots we can see that Science is the skill with the largest relative and absolute difference between the HOO and other occupations.  

```{r fig2, fig.cap='Weighted average skills by HOO vs other', fig.width=9, fig.height=11, dpi=400, out.width="110%"}
ro$skills_plt
```

```{r, fig3, fig.cap="Relative differences", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
ro$diffs_plt1
```

```{r, fig4, fig.cap="Absolute differences", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
ro$diffs_plt2
```

## Mean values and growth rates of skills

An alternative view is to look at the labour market as a whole and make comparisons across time. We calculate the weighted average skills at various points in time, and then we calculate the compound annual growth rate.  This gives us a sense of how important each skill is currently, and how it is changing over time. The following plot shows the weighted average skill scores in 2024 (Figure \@ref(fig:fig5)) and the compound annual growth rate (CAGR) of these weighted average skill scores from 2015 to 2024 (Figure \@ref(fig:fig6)), and from 2025 to 2035 (Figure \@ref(fig:fig7)).  Consistent with the HOO vs. other comparison above, we see that historically Science has been the fastest growing skill, forecast to be replaced by programming over the next ten years.  

```{r, fig5, fig.cap="Weighted average skills 2024", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
ro$crossed_plt1
```

```{r, fig6, fig.cap="Historic change", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
ro$crossed_plt2
```

```{r fig7, fig.cap="Future change", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
ro$crossed_plt3 
```

## Skill Clusters

The nature of work is shifting in British Columbia, driven by technological and demographic changes. Regarding technology, both automation and artificial intelligence are transforming how work is done, while demographic changes include an aging population and the increasing importance of diversity and inclusion in the workplace.  Understanding which occupational skills will be important in the jobs of the future is essential for helping job seekers and employers adapt to the evolving economy and labour market. This information is also valuable for educators and policymakers in designing programs that equip the future workforce with the skills and competencies needed for success over the next ten years. In previous labour market reports, we analyzed the demand for individual skills (2022 LMO) and occupational clusters based on skill profiles (2023 LMO).  

While deriving the occupational clusters (LMO 2023) we discovered that many of the O\*NET skills and competencies are highly correlated: they essentially are duplicate measures of the same underlying skill. This high correlation among skills suggests that they are often used together in similar ways across different occupations.  For example, our analysis shows that the correlation coefficient among the ‘Social’ skills exceeds 0.9. As a result, this report shifts focus from individual skills to clustering skills based on their usage patterns across occupations.

Whereas many clustering techniques are available to group observations (such as the 11 occupational clusters in the 2023 LMO), it is a less common task to cluster variables (the skill measures).  To do so we utilized the R clustering package ClustOfVar, developed by Chavent et al. (2012), and validated our results using another R package, ClustVarLV, developed by Vigneau and Qannari (2003). 

The resulting skill clusters revealed distinct groupings, which we termed Analytical, Managerial, Social, STEM, and Technical skill clusters. These clusters offer a more interpretable and actionable framework for understanding future skills demand.

To validate the use of averaged skill cluster scores, we compared their distribution to the original standardized O\*NET importance ratings for individual skills, as mapped to Canadian occupations. Because cluster scores are derived by averaging multiple individual skills, their distribution is smoother and exhibits less variability, particularly at the extremes(Figure \@ref(fig:fig8)). Applying fixed importance thresholds from the standardized O\*NET scale (e.g., scores of 25 and 50) directly to the cluster scores would risk understating importance due to this smoothing effect. To address this, we realigned the thresholds based on percentile positions in the original distribution:

* We first identified the percentiles in the individual skill distribution corresponding to importance scores of 25 ("Less Important") and 50 ("Important").
* These scores corresponded to approximately the 30th and 60th percentiles, respectively.
* We then located the corresponding 30th and 60th percentiles in the skill cluster distribution, which yielded new threshold values (e.g., ~27.5 and ~46.4).
* The realigned values were used to categorize cluster-level scores into importance tiers, preserving the relative interpretation of skill importance.

This percentile-aligned mapping ensures that the meaning of "Important" remains consistent across individual and averaged scores, offering a more accurate framework for classifying skill clusters.

```{r, fig8, fig.cap="Comparison of skill importance distributions", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
plot_dist
```


Next, we linked the average cluster scores to projected job openings, enabling us to identify which skill clusters are most likely to be in demand over the next decade. Finally, we presented results for all occupations and provided disaggregated insights by grouping occupations according to the Training, Education, Experience, and Responsibilities (TEER) framework.

The analytical skills cluster includes foundational abilities such as active listening, reading comprehension, and writing, which tend to be highly rated across many occupations. It also encompasses cognitive skills like critical thinking, complex problem solving, and judgment and decision making, which are similarly valued across diverse roles. Our findings indicate that analytical skills will be essential for success in 74% of job openings over the next ten years and considered somewhat important in the remaining 26%. This suggests that some level of competency in these skills will be required for all job openings during this period (Figure \@ref(fig:fig9)).

Some level of competency in social and management skills is also required across nearly all occupations. These skills are rated as either important in 63% and 36% of projected job openings, respectively, or somewhat important in 36% and 64%, covering almost all job openings over the next ten years. In contrast, STEM and technical skills are rated as important in less than 5% of future job openings, and somewhat important in only 22% and 16%, respectively (Figure \@ref(fig:fig9)). This suggests that, from a broad labour market perspective, these skill clusters are considered less universally critical than others. However, they are often more specialized and occupation-specific, with their relevance becoming more pronounced when disaggregated by factors such as Training, Education, Experience, and Responsibility (TEER) levels..



```{r, fig9, fig.cap="All Occupations: Percent of Job Openings 2025", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
all_occs_chart
```

Management occupations (TEER 0) typically demand higher levels of expertise and experience, which is reflected in our analysis. Almost all job openings in this category require a broad range of essential skills, including analytical, social, and management skills. In occupations requiring a university degree (TEER 1), analytical skills are rated as important in 99% of future job openings. While social and management skills are also essential for this group, the level of proficiency required is somewhat lower compared to that expected in management-specific roles (Figure \@ref(fig:fig10)).

STEM skills show relatively higher importance in occupations requiring a university degree (TEER 1), and to some extent in TEER 0 occupations, albeit at a lower proficiency level. STEM skills are rated as important in 18% of future job openings that require a university degree, reflecting their relevance in more specialized and knowledge-based occupations (Figure \@ref(fig:fig10)). In contrast, technical skills are more commonly emphasized in occupations classified under TEER 2 to TEER 5, which typically require college/apprenticeship training or a high school diploma or less. While technical skills are rarely rated as "Important", they are still considered moderately important in 20% to 38% of job openings within these categories, highlighting their role in supporting practical, hands-on work across a wide range of occupations (Figure \@ref(fig:fig11)).


```{r, fig10, fig.cap="Management and University", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
plot_upper
```

```{r fig11, fig.cap="College/Apprenticeship and Highschool or less", fig.width=9, fig.height=5.5, dpi=400, out.width="110%"}
plot_lower
```

```{r, echo=FALSE, results='asis'}
#cat(rep("\n\n&nbsp;\n\n", 15), sep = "") #hack to force the table onto next page... can't do pagebreaks with bookdown
```


```{r, echo=FALSE, results='asis'}
# Render the table
library(knitr)
kable(skill_clusters, caption = "Skill Cluster Members", booktabs = TRUE)
```
